Fix a bug where keepalive events only fire once (#2502)
This commit fixes a bug where keepalives and check TTL failure
events only fire once. The fix makes failing events occur in
a loop until the next successful event is observed.

This strategy only works for the lifetime of the backend that is
handling the keepalive. If the backend is taken offline, failure
events will not continue to be emitted. A more general solution is
needed, but this is better than what we have now.

Signed-off-by: Eric Chlebek <eric@sensu.io>