Add handlers to v1.CheckResult
This commit adds the Handlers field to v1.CheckResult, bringing us
in line with 1.x behaviour for the client socket.

The commit also cleans up a bit of the agent package, by avoiding
the use of pointers for concurrency primitives in the Agent type.
This makes it easier to write tests that rely on the Agent, since
people won't forget to initialize mutexes and waitgroups.

Finally, the use of timeproxy in the tests has been adjusted to
use a non-zero unix time.

Signed-off-by: Eric Chlebek <eric@sensu.io>