Add package ringv2
This commit adds the ringv2 package. It replaces most of the
functionality in the ring package.

The ringv2 package is much simpler than the original ring package
and operates on the basis of watchers.

A caveat of the new package is that the smallest interval supported
is 5s, a constraint imposed by etcd.

Signed-off-by: Eric Chlebek <eric@sensu.io>