Split agent subscriptions properly in CLI flags
This commit fixes a bug where agents were sending the wrong
subscriptions in entities, as they would not be split properly
before being sent to the config.

Due to a quirk in how subscriptions are handled, this doesn't
affect any current scheduling mechanism. Nevertheless,
subscriptions must be properly stored in entities, especially for
future work that depends on entity details being correct.

Signed-off-by: Eric Chlebek <eric@sensu.io>