Fix round robin scheduling issue
This commit fixes an issue with round robin scheduling, where
two or more checks that share a subscription are using round robin
scheduling.

Triggers are now stored independently of subscriptions, so there
can be any number of ring triggers for a given ring.

This fixes an issue discovered during QA.

Signed-off-by: Eric Chlebek <eric@sensu.io>