Don't allow metrics to be stored in any case
This commit explicitly nils out metrics when events arrive at the
store, in order to ensure that metrics are never stored in etcd.

Care is taken to not write to the pointer that is given to the
store. To this end, some unrelated existing code in the same
UpdateEvent method has also been modified to be safer.

Signed-off-by: Eric Chlebek <eric@sensu.io>