Fix issue causing global flags to be ignored (#2770)
* Update cobra to v0.0.3

This gives us access to the FParseErrWhitelist field on Command objects.
This field can be used to define errors that can be ignored while
parsing, such as "unknown flag" errors.

* Fix issue causing ignored flags

When Load() is called, some sub-command local flags, such as
--format, are not registered yet and this leads to "unknown flags"
errors being returned by cobra. Such an error can throw off the cobra
parser, leading to all the flags appearing after the offending,
supposedly unknown flag to be ignored.

For now, we just ignore such "unknown flags" errors in order not to
potentially ignore other flags as a side effect. cobra will still store
the name/value of all the flags, even if they are registered later.

Unfortunately, a (rather big) refactor of the CLI is probably the
only way to get around that. Such a refactor would involve completely
reordering the way we define commands, load the configuration files
and override properties with command line flags.

See #2345, #2731.

Signed-off-by: Cyril Cressent <cyril@sensu.io>