Make API return continue token
If there is a continue token, it is returned by the API via a
X-Sensu-Continue header in the HTTP response.

The token is base64url encoded (no padding, see RFC4648) to avoid having
to URL encode '/' and the final null byte that tokens typically use.
This only concerns the REST API; the store methods use the decoded form
of the token.

Signed-off-by: Cyril Cressent <cyril@sensu.io>