Work around keepalive events bug
As described in sensu-go#2465, keepalive events do not always have their
namespace filled in, which can break the construction of the continue
token used for pagination. To accommodate for that, when constructing
the continue token, whevener an event has a namespace of "" we construct
the continue token using its entity's namespace instead.

Even when we fix the root cause of sensu-go#2465, we'll need to keep
this workaround for a while in order to still be able to paginate
through old, existing data.

Signed-off-by: Cyril Cressent <cyril@sensu.io>