Use contexts for cancellation
Use contexts for cancellation through the agent application. This
means threading a context through components that are also used in
the backend, so the backend has been nominally modified as well.

Note that cancellation problems still exist in the backend and will
require a larger, focused effort to resolve.

The Agent Stop() method has been removed, and Agent Run() now
blocks until the agent context is cancelled.

The asset manager now accepts a context, and will stop trying to
retrieve assets when the context is cancelled. This prevents
problems with asset managers never stopping, when assets cannot be
retrieved.

The agent's connection manager now stops trying to reconnect when
a context cancellation is observed.

Agents now shut down cleanly when SIGTERM and SIGINT are received.
All in-progress HTTP connections are given 1s to shut down and are
then terminated forcibly.

Agents still block on SIGTERM and SIGINT, but the longer waits for
shutdown should now be resolved.

Signed-off-by: Eric Chlebek <eric@sensu.io>