Audit and fix all etcd watchers
During a recent QA session, it was discovered that watchers could
fail permanently, but their consumers were using them as if they
could not. This was causing issues such as 100% CPU usage in some
components, and other components would simply fail to update until
service restart.

The problem mostly presents on backends that are using the
--no-embed-etcd mode, but can theoretically also occur on backends
that do embed etcd.

All watchers now get reinstated on a retry loop when they fail, and
all watchers now require that a leader is present, so that watchers
do not get stuck on partitions.

Signed-off-by: Eric Chlebek <eric@sensu.io>
Signed-off-by: Nikki Attea <nikki@sensu.io>