Move Config mock to avoid circular dependencies
Some client authentication tests in cli/client need to use a mock Config
object and import cli/client/testing to use the one defined there.

I need to add types to the cli/client package, which are going to be
used by the mock client defined in cli/client/testing to meet the Client
interface.

That would create a circular dependency because the tests in cli/client
would import cli/client/testing and the mock in cli/client/testing would
import cli/client to use its types.

I'm creating a different mock config that lives in the cli/client/config
package, where I think it should have been created in the first place,
so that the authentication tests can use that instead and not create a
circular dependency between cli/client and cli/client/testing.

Ultimately, the config mock that's in the testing package should be
removed altogether. But that entails a lot of refactoring across the
code base.

Signed-off-by: Cyril Cressent <cyril@sensu.io>