Release 5.6.0 (#2897)
* Create 5.6.0 release branch

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Bugfix for annotations & labels flags (#2881)

Signed-off-by: Simon Plourde <simon@sensu.io>

* Move dashboard source code into its own repository (#2869)

As our project has grown, build targets increased, and time has passed, the ideal of a monorepo housing both the web and backend has simply caused more friction than it has reduced. Today we are moving Sensu Go Web into it's own repository: sensu/web.

Signed-off-by: James Phillips <jamesdphillips@gmail.com>

* [GraphQL] Ensures that all types are registered (#2875)

Types that are not directly referenced by the root Schema type or any of
their children are not immediately registered with the schema. As such to
ensure that ALL types are available we append any that are missing.

Signed-off-by: James Phillips <jamesdphillips@gmail.com>

* Add message bus to tessend (#2886)

Signed-off-by: Nikki Attea <nikki@sensu.io>

* Feature/cluster id hex (#2893)

Signed-off-by: Nikki Attea <nikki@sensu.io>

* Add support for field selectors (#2892)

Signed-off-by: Simon Plourde <simon@sensu.io>

* Proxy api from dashboard daemon (#2885)

Signed-off-by: James Phillips <jamesdphillips@gmail.com>