Update CHANGELOG to describe new `handler` queries in GraphQL schema
This commit adds a line item to the latest changelog that describes a change that allow users to query for a handler and a handler collection from the GraphQL schema. Users can now query for a single handler given a namespace and name from the root GraphQL `query` type. Users can also query for a collection of handlers from the `namespace` type.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>