Allow users to query for Handlers from the GraphQL endpoint (#2945)
* Add `Handler` fields and types to the GraphQL Schema

This commit adds a `handler` field to the root GraphQL schema that allows users to query for a `Handler` given a namespace and a handler name. We've also added the `HanderConnection` and `HandlerEdge` types to the schema that will allow users to retrieve a collection of `Handlers`.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>

* Update the `namespace` GraphQL type to inlcude `handlers` connection field.

This commit adds a `handers` field to the `Namespace` type that allows users to query for a collection of `Handlers` given the usual connection parameters (offset, limit, orderBy, and filter). This field will be used to display a list of handers in the Web UI.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>

* Add a resolver for `handers` field in the `namespace` type.

This commit updates namespace implementation to inlcude resolver for `handlers` query which allows users to retrieve a collection of handlers for the selected `namespace`. To suppor this change, we've created sorting functions for the `handler` type that allows users to sort handers by name (asc and desc).

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>

* Update CHANGELOG to describe new `handler` queries in GraphQL schema

This commit adds a line item to the latest changelog that describes a change that allow users to query for a handler and a handler collection from the GraphQL schema. Users can now query for a single handler given a namespace and name from the root GraphQL `query` type. Users can also query for a collection of handlers from the `namespace` type.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>