Fix Namespace and User pagination (#2960)
A refactor had broken the computation of their continue token. It was
easy to overlook because both Namespace and User are slight snowflakes
in the sense that they are almost like all the other resource types,
except that their object metadata is always empty. This exception makes
it so that we need specialized code to compute their continue token.

Signed-off-by: Cyril Cressent <cyril@sensu.io>