Refactor eventd (#3011)
This commit moves a significant portion of event update logic to
the core/v2 package, and another portion to the store package.

Now, the store/etcd.UpdateEvent method does all of the necessary
logic to update an event in-place, and returns the updated event
and also the previous event.

Flap detection logic has been moved into the core/v2 package, and
has been incorporated into the Check.MergeWith method.

This enables a better data access pattern for stores that can
handle event updates in a single transaction.

As an incidental bug fix, check state and check total state change
have been corrected. The previous algorithm used the oldest history
item when doing a comparison with the newest event, rather than the
most recent history item. Additionally, the algorithm looped over
the history in the wrong order before. It now loops over the
history in descending order.

Signed-off-by: Eric Chlebek <eric@sensu.io>