Reduce CPU usage in proxy entity filtering
This commit uses a batching strategy for filtering proxy entities,
so that all of the proxy entities that are to be filtered for a
given check are done so with a single Javascript invocation,
instead of M*N invocations. This should dramatically reduce CPU
usage for scheduling proxy checks.

Below is a benchmark of the previous filtering mechanism compared
with the new one.

BenchmarkEvaluate1000-4                10 183369379 ns/op 174297743 B/op  1581652 allocs/op
BenchmarkEvaluateEntityFilters1000-4  100  14916854 ns/op 6278989 B/op	  166571  allocs/op

The new mechanism is over 10 times faster at filtering 1000 proxy
entities, uses 10 times fewer allocations, and 6 MB instead of
174 MB of memory.

The only downside to this approach is that it's no longer possible
to detect syntax errors versus semantic errors in javascript
filters, so all errors are logged at debug level.

Signed-off-by: Eric Chlebek <eric@sensu.io>