[Graphql] Simple autosuggestion (#3028)
Adds a field to the GraphQL service that returns a set of values associated with a given field, type and namespace. Enabling add initial rudimentary autosuggestion features to our clients.

Ref

The 'ref' is an identifier used to uniquely reference a resource in the system as well as a field on said resource. Refs take the form: :group/:version/:type/:field. (Within reason,) :group/:version should always match the package, :type should match a name registered in the typemap, and the :field components should be snake_case. Further, the field segment may be a path in and of it's own, eg. metadata/name would refer to the name field nested inside a resource's metadata.

The following are valid examples of refs:

  core/v2/asset/metadata/name
  core/v2/asset/metadata/labels
  core/v2/asset/metadata/labels/region
  core/v2/check_config/subscriptions
  core/v2/check_config/command
  core/v2/check_config/timeout
  core/v2/entity/system/os
  core/v2/entity/system/platform
  core/v2/filter/metadata/name
  core/v2/handler/command
  core/v2/hook_config/command
  core/v2/mutator/command
  core/v2/mutator/timeout
  core/v2/silenced/creator

A Star Trek Villian

The 'q' argument can be used by the consumer to filtering out the values that do not contain the given sequence of characters. Matches are case-insensitive and in this initial implementation no other analysis/scoring/fuzzy matching occurs.

Usage

  query HandlerSuggestions {
    # fetch the first five handler names; values are sorted alphanumerically
    suggest(ref: "core/v2/handler/metadata/name", namespace: "default", limit: 5, order: ALPHA_DESC) {
      values
    }
  }
  
  query LabelSuggestions {
    # fetch the ten most popular label names among checks
    suggest(ref: "core/v2/check_config/metadata/labels", namespace: "default", order: FREQUENCY) {
      values
    }
  }
  
  query LabelValueSuggestions {
    # among checks, fetch the first ten values under the label "region". Sorted alphanumerically.
    suggest(ref: "core/v2/check_config/metadata/labels/region", namespace: "default", order: ALPHA_ASC) {
      values
    }
  }
  
  query CheckCommandSuggestions {
    # among checks, fetch the first ten commands that contain the string "echo".
    suggest(ref: "core/v2/check_config/command", namespace: "default", q: "echo", order: ALPHA_DESC) {
      values
    }
  }
  
  query MutatorTimeoutSuggestions {
    # among mutators, fetch the five most frequently used values for timeout.
    # note that all items in the result set are strings regardless of the original type.
    suggest(ref: "core/v2/mutator/timeout", namespace: "default", order: FREQUENCY, limit: 5) {
      values
    }
  }

Signed-off-by: James Phillips <jamesdphillips@gmail.com>