Add --disable-assets flag to the agent
This commit adds an agent command line flag, and ability via the
agent config file, to disable assets. In the event that an agent
is instructed to execute a check that requires an asset, the agent
will respond with a status of 3, and a message indicating that the
agent could not execute the check, because assets are disabled.

Also added is the ability to use the null device as a cache
directory. When the null device (/dev/null on nix, NUL on Windows)
is used, the agent will not attempt to create an event queue.

If --disable-assets is not set, and the cache directory is set to
the null device, then an error will occur.

These changes make the agent require no storage to operate, at the
cost of the asset and durable events API features.

Signed-off-by: Eric Chlebek <eric@sensu.io>