Add an `upsert` parameter to the `PutWrapped` mutation (#3049)
* [Web] Update `PutWrapped` mutation definition

This change updates the `PutWrapped` mutation to add an additional `upsert` argument to the mutation. The addition of the `upsert` argument fixes a bug where resources would be overwritten (upserted) when the users of the mutation would attempt to create a resource using the same name as an existing resource. This happened because the `PutWrapped` mutation would call the PUT endpoint under the hood. Since we don't always want this behaviour, we're adding the `upsert` param (boolean flag) that will conditionally call the POST endpoint if `upsert === false`. By default, `upsert` is set to `true` to maintain the existing behaviour of the mutation for backwards compatibility.

This change updates the mutations schema to reflect the addition of the `upsert` argument in the `PutWrapped` mutation.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>

* Update `PutWrapped` mutation to use `upsert` param

This change updates the `PutWrapped` mutation resolver in order to conditionally call the resource POST endpoint if the user of the mutation does not want the mutation to upsert the resource they are passing as the `raw` argument of the mutation. Previously, the mutation would always call the resource POST endpoint, which overwrites (upserts) the resource even if the resource already exists. Since users of the mutation don't always want this behaviour, we're updating the implementation of the `PutWrapped` mutation resolver to conditionally call the POST endpoint if the `upsert` parameter is `false`. By default, the `upsert` parameter is `true` in order to maintain the original behaviour of the mutation (for backwards compatibility).

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>