[GraphQL] Queries for 'Event Filter' Resources (#3085)
* [GraphQL] Update schema with event filter queries

This change updates the GraphQL Schema with the ability to query for Event Filter resources. Now queries for a single Event Filter can be made from the root `query` node via the `eventFilter` field. Or a list of Event Filters can be queried from the `namespace` node via the `eventFilters` field.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>

* [GraphQL] Create resolvers for Event Filters

This change creates resolvers for the newly added Event Filter queries in the GraphQL Schema.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>