Refactor the storage requirements for Authorizer
Using the Authorizer in the enterprise codebase is cumbersome, due
to the large API surface that must be mocked. Because of an
implementation peculiarity in our mocking library, we cannot use
the sensu-go mock store, and must therefore recreate our mocks
as necessary.

By specifying the exact API of consumers through interfaces, this
process becomes much easier.

Signed-off-by: Eric Chlebek <eric@sensu.io>