Fix a bug in proxy check scheduling
This commit fixes a bug where proxy check schedulers can be shut
down and recreated too often. In some cases this can lead to checks
not being executed frequently enough.

The problem is caused by entity updates being treated as cause for
shutting down schedulers, when in fact, it is only necessary to do
so in the rare circumstance that entity attributes change, leading
to a different cardinality of round robin ring.

To avoid this happening, before cancelling the existing scheduler,
the cardinality of the request is inspected, and if unchanged, the
existing scheduler is left alone.

Signed-off-by: Eric Chlebek <eric@sensu.io>