Fix a bug where sensuctl dump errors are silently discarded
This commit reworks sensuctl dump to use the API client instead of
shelling out to sensuctl. This has the benefit of allowing the
tool to ignore some HTTP errors, but not others.

Additionally, resources are now dynamically resolved, so new types
of resources don't need to be explicitly added to the tool.

Signed-off-by: Eric Chlebek <eric@sensu.io>