Use the embedded client when etcd is embedded (#3264)
This commit makes Sensu use an embedded client for etcd requests,
instead of connecing localhost. This has several useful effects:
  * Avoids the network stack for etcd requests, for embedded etcd
  * Avoids gRPC serialization, for embedded etcd
  * Simplifies integration testing

The biggest win here is that we can have the integration test etcd
servers listen on ':0', which assigns an arbitrary port instead of
one that we have to pick. This should let the OS pick a port
instead of us, which solves a long-standing CI bug where our tests
could sometimes pick the same port to listen on in parallel,
causing failure.

There should also be some performance wins from avoiding the
network stack, but this is yet to be verified.

Signed-off-by: Eric Chlebek <eric@sensu.io>