Fix 'flag accessed but not defined' error in asset outdated (#3298)
Missing flag declarations were causing the command to always fail. A
simple test has been added to check that the command runs properly and
just returns a tabular header when there are no assets. Adding more
tests will require refactoring the Bonsai client out of the command so
that it can be mocked.

Signed-off-by: Cyril Cressent <cyril@sensu.io>