Stop and restart daemons on error
When the backend error group receives an error, the sensu-backend
now stops and restarts all of its internal daemons.

Previously, if a daemon wrote an error to its error channel, the
entire service would stop running.

This change should make it easier to ensure that the system does
not end up in a half-working state. If an daemon writes an error
to its error channel, the rest of the daemons will stop and
restart.

Signed-off-by: Eric Chlebek <eric@sensu.io>