Return structured errors from the store
This commit modifies the store to return structured errors for
each error condition. The intent is that the user of the store
will always be able to determine the reason for store failure.

If ErrInternal is returned, the user can assume that the error is
not related to the request, but rather is an error emanating from
the underlying database.

The intent is that consumers will be able to distinguish between
request format errors, and errors that would represent a loss of
service.

Signed-off-by: Eric Chlebek <eric@sensu.io>