Emit fatal errors on store ErrInternal
This commit modifies agentd, eventd, keepalived, and pipelined to
emit fatal errors to the backend supervisor, when they receive
ErrInternal on any store method.

This will result in the shutdown and restart of all daemons, which
will give the system a "clean slate" to start from, and will also
sever all existing agent connections.

If the backend is experiencing trouble and can't use the store,
this gives agents a chance to connect to other systems that might
be able to offer service.

Signed-off-by: Eric Chlebek <eric@sensu.io>