Simplify the resource cache, fix memory leak
This commit fixes a memory leak in the resource cache by making it
avoid using etcd watchers. For reasons that are not clear, the
previous implementation was causing a memory leak under load.

It's unknown if the memory leak was a result of improperly using
the etcd client API, a bug in the etcd client API, or caused by
sensu-go in some other way.

Signed-off-by: Eric Chlebek <eric@sensu.io>