Restart backend daemons on store failure (#3446)
* Stop and restart daemons on error

When the backend error group receives an error, the sensu-backend
now stops and restarts all of its internal daemons.

Previously, if a daemon wrote an error to its error channel, the
entire service would stop running.

This change should make it easier to ensure that the system does
not end up in a half-working state. If an daemon writes an error
to its error channel, the rest of the daemons will stop and
restart.

* Return structured errors from the store

This commit modifies the store to return structured errors for
each error condition. The intent is that the user of the store
will always be able to determine the reason for store failure.

If ErrInternal is returned, the user can assume that the error is
not related to the request, but rather is an error emanating from
the underlying database.

The intent is that consumers will be able to distinguish between
request format errors, and errors that would represent a loss of
service.

* Emit fatal errors on store ErrInternal

This commit modifies agentd, eventd, keepalived, and pipelined to
emit fatal errors to the backend supervisor, when they receive
ErrInternal on any store method.

This will result in the shutdown and restart of all daemons, which
will give the system a "clean slate" to start from, and will also
sever all existing agent connections.

If the backend is experiencing trouble and can't use the store,
this gives agents a chance to connect to other systems that might
be able to offer service.

* Refactor pipelined

This commit moves the pipeline business logic out of pipelined and
into a new package called pipeline.

The intent is that future work will use the pipeline directly.

* Terminate sessions when ErrInternal is seen
* Use the retry package for restarting the daemons
* Add store timeouts to several daemons
* Fix a bug where tessend did not completely stop

* Crash if agentd and dashboardd can't listen

When agentd and dashboardd listeners return an error, there is
likely a configuration problem. Establish their listeners in a
blocking manner, so that Start() can return the error.

* Fix agent stop dance

Make sure the agent stop blocks until the ring item is removed.
Use a timeout for removing the ring item as well.

* Get rid of the done channel

Makes the backend stop properly with CTRL+C

Signed-off-by: Eric Chlebek <eric@sensu.io>