Pass an initialization function to start cmd (#3567)
This commit changes the way the sensu-backend starts. The call
to Run has been replaced with a call to RunWithInitializer. This
allows the commerical backend to pass its initializer to the OSS
start command.

Without this change, the commerical backend could lose its
enterprise-specific HTTP routes when the backend experienced an
internal restart.

Signed-off-by: Eric Chlebek <eric@sensu.io>