Add agent platform discovery
This commit adds features that populate the LibCType, VMSystem,
VMRole, and CloudProvider fields in Entity.System.

LibCType, VMSystem, and VMRole are always populated on agent
startup.

CloudProvider is populated if the --detect-cloud-provider flag is
set. This feature can result in several HTTP requests or DNS
lookups being performed, so it may not be appropriate for all
environments.

There are no automated tests for cloud provider discovery yet.
Since the function requires an attempt to connect to a link-local
IP address, it may be difficult to achieve this.

Signed-off-by: Eric Chlebek <eric@sensu.io>