Don't use the etcd embedded client (#3606)
It seems that in some circumstances, the etcd embedded client can
cause nil panics. For instance, when the embedded etcd is shut
down, the client cannot be safely used anymore.

Since this is an unacceptable state of affairs, we will revert to
using the gRPC client, which seems to be much more reliable.

The embedded client is better for integration tests, because it
does not require the client to know about which ports the etcd
service has bound to. We'll continue to use it there.

Signed-off-by: Eric Chlebek <eric@sensu.io>