Fix a bug that caused sensu-backend restart
This commit fixes a bug that caused sensu-backend to restart
whenever an agent terminates its connection. The bug was
introduced in 572b4b5facc4a4f7b310dcf0eb34ef711e9583e5, but went
unnoticed for the most part, due to the use of the embedded etcd
client, which appears to have different cancellation semantics than
the gRPC client.

Now that the gRPC client has been put back into place, this error
was exposed and made reproducible.

The nature of the error is that when RBAC authorization returned
an error, agentd would propagate the error back to its supervisor
in all cases, leading to an internal backend restart.

When a session's context is canceled, whether due to its client
disconnecting or otherwise, agentd should not propagate any
errors back to the backend supervisor.

Signed-off-by: Eric Chlebek <eric@sensu.io>