Don't service dashboard requests from API clients
This commit changes the dashboard service to write a status of 400
and an error message explaining that the user should not try to
connect to it with an API client.

For the purpose of this change, it is assumed that for the most
part, API clients will either be sensuctl or curl.

This should lead to less confusion, as users are prone to
connecting to the dashboard port with sensuctl or curl, and tend to
get confused when the service produces different results than what
is documented for the apid service running on a different port.

Signed-off-by: Eric Chlebek <eric@sensu.io>