Keepalives now work correctly when using postgres
This commit fixes an issue where keepalives would not always fire
correctly when using the postgres store. It seems that this
regression was introduced several releases ago, but it was not
always apparent for all users.

The keepalive mechanism was always consulting etcd when retrieving
events, and as such, when postgres was enabled, keepalives could be
incorrectly passing.

Unfortunately, due to the current architecture of sensu-go and
sensu-enterprise-go, it is not possible to write a unit test that
proves this commit fixes the issue. A test will need to be added
to the sensu-enterprise-go repo instead.

An unrelated unit test has been added that shores up testing of
check history generation; this has been done to maintain parity
between sensu-go and sensu-enterprise-go for tests.

Signed-off-by: Eric Chlebek <eric@sensu.io>