Windows agent: fix flag issues, add log rotation
This commit makes signifcant fixes and improvements to the Windows
sensu-agent, including log rotation features and improvements to
the way the service installer handlers its arguments.

The service installer now supports all of the arguments that the
regular sensu-agent start command does, and will remember those
arguments when launching the application.

Log rotation is now supported, with support for configurable
log file sizes, and reaping old log files. On rotate, logs are
zipped.

Signed-off-by: Eric Chlebek <eric@sensu.io>