Token substitution for assets (#3720)
This commits adds support for token substitution in runtime assets for the following resources:

- checks (agent)
- hooks (agent)
- handlers (backend)
- mutators (backend)

Essentially, it relies on the filtered asset manager to perform the token substitution before retrieving it; this manager already has access to the entity (whether it's the agent or the backend) so it was easy to add.

In addition, it refactors the token substitution logic into its own package. 

Also, it removes the asset validation rules around the URL; we can now rely on token substitution at runtime therefore it's impossible to validate these URLs. Additionally, I modified the create & update generic handlers so they return the underlying error to sensuctl create and other clients.