Support agent EntityConfig updates from backend
This commit adds support for receiving updates about entity
configuration from sensu-backend to sensu-agent. When a user
updates an entity's configuration, the backend will send a message
to the corresponding connected agent informing it that the entity
configuration has changed. The agent will correspondingly update
its cached version of the entity configuration.

A small unrelated change was also made in the interests of
completeness for Sensu 6.0. The sensu-backend init tool will now
also upgrade the cluster to the latest version, so that new
clusters will start out at the correct version.

Signed-off-by: Eric Chlebek <eric@sensu.io>