Add Entity state update exception (#3911)
We now have 2 code paths when updating entities: one for proxy entities and another for all    
other types of entities. We had to make that distinction because Entity  
is still the public API to interact with entities, even though internally
we use the storev2 EntityConfig/EntityState split.                       
                                                                         
The consequence was that updating an Entity could alter its state,       
something we don't really want unless that entity is a proxy entity.     
                                                                         
See sensu-go#3896.                                                       

Signed-off-by: Cyril Cressent <cyril@sensu.io>