Use NamespaceStore in NamespaceClient.Delete()
We don't use the generic client and store in Delete() because there is
some special logic that applies to namespace deletion, namely the fact
that we don't want to delete namespace objects as if they were
independent objects: we want to make sure that a namespace is logically
"empty" before we remove it for good.

Since we don't have a good abstraction for these types of custom logic
in the generic client and store yet, we reach straight to the
NamespaceStore, which already has this logic implemented.

We can revisit this approach in the future, if/when we rework the
generic objects to better support "per type custom logic".

Signed-off-by: Cyril Cressent <cyril@sensu.io>