Remove roles from namespace emptiness definition
While it's technically incorrect to call a namespace that still has
roles "inside it" empty, we sort of need to do that for now to unblock
deleting namespaces and their new, implicit pipeline role and binding.

Checking that core objects such as checks, handlers, filters, ... are
not in a namespace anymore is the really important part.

Signed-off-by: Cyril Cressent <cyril@sensu.io>