Reorder operations in DeleteNamespace()
We try to delete the namespace before deleting the associated pipeline
role and binding so that in case the namespace deletion fails because
the namespace is considered non-empty, we don't have to revert the
role/binding deletion.

Also, failing to delete those pipeline role/binding objects only causes
a warning, since it's technically fine to leave them behind.

Signed-off-by: Cyril Cressent <cyril@sensu.io>