Don't wait for graceful termination on agent halt
This commit adds a timeout to agent termination so that graceful
shutdown can't block indefinitely. It seems there is a bug in
sensu-agent that will cause an infinite loop between the backend and
agent to occur.

This only papers over the real problem, but it will at least improve the
situation somewhat until the larger issue can be addressed.

Signed-off-by: Eric Chlebek <eric@sensu.io>